Moved state machine from text sensor into its own namespace
Also merged state machine base logic and component
