+++ components/state_machine/automation.h
@@ -2,121 +2,123 @@
 
 #include "esphome/core/component.h"
 #include "esphome/core/automation.h"
-#include "state_machine_text_sensor.h"
+#include "state_machine.h"
 
 namespace esphome
 {
-
-  class StateMachineOnEnterTrigger : public Trigger<>
+  namespace state_machine
   {
-  public:
-    StateMachineOnEnterTrigger(StateMachineTextSensor *text_sensor, std::string state)
+
+    class StateMachineOnEnterTrigger : public Trigger<>
     {
-      text_sensor->add_on_transition_callback(
-          [this, state](StateTransition transition)
-          {
-            this->stop_action(); // stop any previous running actions
-            if (transition.to_state == state)
+    public:
+      StateMachineOnEnterTrigger(StateMachineComponent *state_machine, std::string state)
+      {
+        state_machine->add_on_transition_callback(
+            [this, state](StateTransition transition)
             {
-              this->trigger();
-            }
-          });
-    }
-  };
+              this->stop_action(); // stop any previous running actions
+              if (transition.to_state == state)
+              {
+                this->trigger();
+              }
+            });
+      }
+    };
 
-  class StateMachineOnLeaveTrigger : public Trigger<>
-  {
-  public:
-    StateMachineOnLeaveTrigger(StateMachineTextSensor *text_sensor, std::string state)
+    class StateMachineOnLeaveTrigger : public Trigger<>
     {
-      text_sensor->add_on_transition_callback(
-          [this, state](StateTransition transition)
-          {
-            this->stop_action(); // stop any previous running actions
-            if (transition.from_state == state)
+    public:
+      StateMachineOnLeaveTrigger(StateMachineComponent *state_machine, std::string state)
+      {
+        state_machine->add_on_transition_callback(
+            [this, state](StateTransition transition)
             {
-              this->trigger();
-            }
-          });
-    }
-  };
+              this->stop_action(); // stop any previous running actions
+              if (transition.from_state == state)
+              {
+                this->trigger();
+              }
+            });
+      }
+    };
 
-  class StateMachineTransitionActionTrigger : public Trigger<>
-  {
-  public:StateMachineTransitionActionTrigger(StateMachineTextSensor *text_sensor, StateTransition for_transition)
+    class StateMachineTransitionActionTrigger : public Trigger<>
     {
-      text_sensor->add_on_transition_callback(
-          [this, for_transition](StateTransition transition)
-          {
-            this->stop_action(); // stop any previous running actions
-            if (transition.from_state == for_transition.from_state 
-              && transition.input == for_transition.input 
-              && transition.to_state == for_transition.to_state)
+    public:
+      StateMachineTransitionActionTrigger(StateMachineComponent *state_machine, StateTransition for_transition)
+      {
+        state_machine->add_on_transition_callback(
+            [this, for_transition](StateTransition transition)
             {
-              this->trigger();
-            }
-          });
-    }
-  };
+              this->stop_action(); // stop any previous running actions
+              if (transition.from_state == for_transition.from_state && transition.input == for_transition.input && transition.to_state == for_transition.to_state)
+              {
+                this->trigger();
+              }
+            });
+      }
+    };
 
-  class StateMachineInputActionTrigger : public Trigger<>
-  {
-  public:
-    StateMachineInputActionTrigger(StateMachineTextSensor *text_sensor, std::string input)
+    class StateMachineInputActionTrigger : public Trigger<>
     {
-      text_sensor->add_on_transition_callback(
-          [this, input](StateTransition transition)
-          {
-            this->stop_action(); // stop any previous running actions
-            if (transition.input == input)
+    public:
+      StateMachineInputActionTrigger(StateMachineComponent *state_machine, std::string input)
+      {
+        state_machine->add_on_transition_callback(
+            [this, input](StateTransition transition)
             {
-              this->trigger();
-            }
-          });
-    }
-  };
+              this->stop_action(); // stop any previous running actions
+              if (transition.input == input)
+              {
+                this->trigger();
+              }
+            });
+      }
+    };
 
-  template <typename... Ts>
-  class StateMachineTransitionAction : public Action<Ts...>, public Parented<StateMachineTextSensor>
-  {
-  public:
-    StateMachineTransitionAction(std::string input)
+    template <typename... Ts>
+    class StateMachineTransitionAction : public Action<Ts...>, public Parented<StateMachineComponent>
     {
-      this->input_ = input;
-    }
+    public:
+      StateMachineTransitionAction(std::string input)
+      {
+        this->input_ = input;
+      }
 
-    void play(Ts... x) override { this->parent_->transition(this->input_); }
+      void play(Ts... x) override { this->parent_->transition(this->input_); }
 
-  protected:
-    std::string input_;
-  };
+    protected:
+      std::string input_;
+    };
 
-  template <typename... Ts>
-  class StateMachineTransitionCondition : public Condition<Ts...>
-  {
-  public:
-    explicit StateMachineTransitionCondition(StateMachineTextSensor *parent) : parent_(parent) {}
-
-    TEMPLATABLE_VALUE(std::string, from_state)
-    TEMPLATABLE_VALUE(std::string, input)
-    TEMPLATABLE_VALUE(std::string, to_state)
-
-    bool check(Ts... x) override
+    template <typename... Ts>
+    class StateMachineTransitionCondition : public Condition<Ts...>
     {
-      if (!this->parent_->last_transition.has_value())
-        return false;
-      StateTransition transition = this->parent_->last_transition.value();
-      if (this->from_state_.has_value() && this->from_state_.value(x...) != transition.from_state)
-        return false;
-      if (this->input_.has_value() && this->input_.value(x...) != transition.input)
-        return false;
-      if (this->to_state_.has_value() && this->to_state_.value(x...) != transition.to_state)
-        return false;
-      return true;
-    }
+    public:
+      explicit StateMachineTransitionCondition(StateMachineComponent *parent) : parent_(parent) {}
 
-  protected:
-    StateMachineTextSensor *parent_;
-  };
+      TEMPLATABLE_VALUE(std::string, from_state)
+      TEMPLATABLE_VALUE(std::string, input)
+      TEMPLATABLE_VALUE(std::string, to_state)
 
+      bool check(Ts... x) override
+      {
+        if (!this->parent_->last_transition().has_value())
+          return false;
+        StateTransition transition = this->parent_->last_transition().value();
+        if (this->from_state_.has_value() && this->from_state_.value(x...) != transition.from_state)
+          return false;
+        if (this->input_.has_value() && this->input_.value(x...) != transition.input)
+          return false;
+        if (this->to_state_.has_value() && this->to_state_.value(x...) != transition.to_state)
+          return false;
+        return true;
+      }
+
+    protected:
+      StateMachineComponent *parent_;
+    };
+
+  } // namespace state_machine
 } // namespace esphome
