+++ dimmable_light_example.yaml
@@ -6,12 +6,11 @@ esphome:
 logger:
   level: DEBUG
 
-external_components:  
+external_components:
   - source: components
 
-text_sensor:
-  - platform: state_machine
-    name: Light Brightness State Machine
+state_machine:
+  - name: Light Brightness State Machine
     initial_state: "OFF"
     states:
       - name: "OFF"
