+++ toggle_example.yaml
@@ -6,12 +6,11 @@ esphome:
 logger:
   level: DEBUG
 
-external_components:  
+external_components:
   - source: components
 
-text_sensor:
-  - platform: state_machine
-    name: On/Off Toggle State Machine
+state_machine:
+  - name: On/Off Toggle State Machine
     states:
       - name: "OFF"
         on_enter:
